src/xfaas/requirements-xfaas.txt
boto3==1.26.137
azure-functions==1.11.2
azure-storage-blob==12.16.0
google-cloud-functions==1.13.0
google-cloud-storage==2.9.0
qiskit==0.43.1
amazon-braket-sdk==1.53.0
asyncio
requests==2.31.0
numpy==1.24.3